Moving route paths
