+++ quesma/quesma/dual_write_proxy_v2.go
@@ -16,6 +16,7 @@ import (
 	"quesma/end_user_errors"
 	"quesma/feature"
 	"quesma/frontend_connectors"
+	"quesma/frontend_connectors/routes"
 	"quesma/ingest"
 	"quesma/logger"
 	"quesma/queryparser"
@@ -24,7 +25,6 @@ import (
 	"quesma/quesma/gzip"
 	"quesma/quesma/mux"
 	"quesma/quesma/recovery"
-	"quesma/quesma/routes"
 	"quesma/quesma/types"
 	"quesma/quesma/ui"
 	"quesma/schema"
