+++ quesma/quesma/router_test.go
@@ -7,10 +7,10 @@ import (
 	"github.com/stretchr/testify/assert"
 	"quesma/clickhouse"
 	"quesma/frontend_connectors"
+	"quesma/frontend_connectors/routes"
 	"quesma/ingest"
 	"quesma/quesma/config"
 	"quesma/quesma/mux"
-	"quesma/quesma/routes"
 	"quesma/quesma/ui"
 	"quesma/schema"
 	"quesma/table_resolver"
