+++ quesma/quesma/router_v2.go
@@ -10,6 +10,7 @@ import (
 	"quesma/clickhouse"
 	"quesma/elasticsearch"
 	"quesma/frontend_connectors"
+	"quesma/frontend_connectors/routes"
 	"quesma/ingest"
 	"quesma/logger"
 	"quesma/queryparser"
@@ -21,7 +22,6 @@ import (
 	"quesma/quesma/functionality/resolve"
 	"quesma/quesma/functionality/terms_enum"
 	"quesma/quesma/mux"
-	"quesma/quesma/routes"
 	"quesma/quesma/types"
 	"quesma/quesma/ui"
 	"quesma/schema"
